requirements.txt
streamlit
plotly
pandas
yfinance
vaderSentiment
python-dotenv
requests
